Splitting classes
