+++ src/components/common/math/MathTextBox/MathTextBox.module.sass
@@ -13,44 +13,3 @@
     font-size: 125%
     line-height: 1
     font-style: italic
-
-.completeContainer
-  display: flex
-  flex-direction: column
-  border-radius: 5px
-
-.variableContainer
-  display: flex
-  justify-content: flex-end
-  margin-top: 20px
-
-.variableTextBox
-  margin-right: 20px
-
-.variable
-  display: flex
-  align-items: center
-  
-  /*div
-  /*  min-width: 50px
-
-  /*span
-  /*  font-size: 0.9em
-
-  /*var
-  /*  font-size: 0.9em
-
-.item
-  margin: 0 5px
-
-.varIcon
-  align-self: center
-  cursor: pointer
-
-.removeVarIcon
-  margin-right: 10px
-
-.addVarIcon
-
-.mathText
-  margin: 0
