+++ src/components/common/math/MathTextBoxWithVariables/MathTextBoxWithVariables.module.sass
@@ -1,0 +1,21 @@
+.container
+  display: flex
+  flex-direction: column
+  border-radius: 5px
+
+.variableContainer
+  display: flex
+  justify-content: flex-end
+  margin-top: 20px
+
+.variableTextBox
+  margin-right: 20px
+
+.varIcon
+  align-self: center
+  cursor: pointer
+
+.removeVarIcon
+  margin-right: 10px
+
+.addVarIcon
