+++ src/components/common/math/MathTextBoxWithVariables/index.js
@@ -1,0 +1,3 @@
+import MathTextBoxWithVariables from './MathTextBoxWithVariables';
+
+export default MathTextBoxWithVariables;
