+++ src/components/common/math/VariableTextBox/VariableTextBox.module.sass
@@ -1,0 +1,18 @@
+.variable
+  display: flex
+  align-items: center
+  
+  /*div
+  /*  min-width: 50px
+
+  /*span
+  /*  font-size: 0.9em
+
+  /*var
+  /*  font-size: 0.9em
+
+.item
+  margin: 0 5px
+
+.mathText
+  margin: 0
