+++ src/components/common/math/VariableTextBox/index.js
@@ -1,0 +1,3 @@
+import VariableTextBox from './VariableTextBox';
+
+export default VariableTextBox;
